Refactor organization page and error handling
The changes involve an overall method handling of the fetchOrganizationAccess function. Instead of having a dedicated methodology, we have incorporated that directly into the OrganizationPageProps function and data request to the server is now done directly in the same scope. This simplifies our code and optimizes error handling. If the page fails to fetch organization access, it returns NotFound component. This modification removes unnecessary code complexity and promotes easier code maintenance.