(fix|COS-488): Implement error handling for inaccessible organization details page (#4110)
* (fix|COS-488): Implement error handling for inaccessible organization on details page

Enhanced the application's flow by adding a mechanism that verifies the visibility and existence of an organization prior to generating the details page. This was achieved through the following updates:

- Implemented a new fetch function that leverages GraphQL to authenticate if a user has the required access to an organization.
- Integrated this fetch function into the main page logic.
- Introduced a new `NotFound` component, which is displayed anytime a user attempts to access an organization that is not accessible or does not exist.

This solution ensures a robust application experience by preventing errors and gracefully managing cases where users try to view unavailable or restricted organizations.

* Refactor organization page and error handling

The changes involve an overall method handling of the fetchOrganizationAccess function. Instead of having a dedicated methodology, we have incorporated that directly into the OrganizationPageProps function and data request to the server is now done directly in the same scope. This simplifies our code and optimizes error handling. If the page fails to fetch organization access, it returns NotFound component. This modification removes unnecessary code complexity and promotes easier code maintenance.