fix(cos-462, cos-508): Preview modal + ui improvements (#4120)
* WIP: "Refactor ViewInExternalAppButton, removing duplication

Created a new ViewInExternalAppButton component that abstracts the shared functionality of ViewInIntercomButton and ViewInSlackButton. This leads to code that is easier to maintain and less prone to bugs. Also updated the places where these old components were used with the new ViewInExternalAppButton component."

* Add IssuePreviewModal and restructure timeline issue card

Made major enhancements to the issue card functionality in the timeline view. Apart from various refinements to the UI and layout, significant changes include:

- Added a new AlertSquare SVG icon in the media folder
- Created a new dynamic PriorityBadge component to display issue priority visually
- Built the IssuePreviewModal to show issue details upon clicking the issue card. This involved a major restructure of existing code in IssuePreviewModal.tsx and improvement of related functionalities. Call to this is currently mocked.
- Added mock data in IssueStub for testing
- Integrated the IssuePreviewModal in the TimelineEventPreviewModal
- Included adjustment in styles in IssueCommentCard for better visual aesthetics

* (fix|COS-508): Issues ui polish

* Added Zendesk logo, adjusted styles, and expanded issue data

Introduced Zendesk's logo for use in the ViewInExternalAppButton component. Several styles were adjusted, including zIndex in the IssueStub style and cursor behaviour in IssueCommentCard.