Change 'does not exist' to 'cannot be found' in organization details NotFound component
This commit changes the error message text displayed when an organization is not found. The previous wording implied that the organization didn't exist at all which is not always the case. The new wording, 'this organization cannot be found', is more accurate as it encompasses all scenarios like the organization belonging to different tenant etc.