Added empty state in email compose when Google OAuth is toggled off (#4177)
* Added empty state in email compose when Google OAuth is toggled off

Created new compose email container that checks if the user has permissions to send an email via Gmail. If the user doesn't have permission, a message will appear prompting them to allow the application to send emails on their behalf. This ensures that users are aware of the application's behavior, giving them better control and improving overall user experience.

* feat: add missing permisions prompt for email composers

* chore: use isGoogleActive from global cache

* build: bypass test steps (temporary)

---------

Co-authored-by: Alex Calinica <acalinica@openline.ai>