(fix|cos-619): Fix for scrollbar overflowing About panel modals
Modify disclosure hook usage for RenewalForecast and RenewalLikelihood
Updated useDisclosure hook, moved it to an upper scope in the AccountPanel component to fix an issue with scrolling. The issue was that scrollbars were overlapping with renewal modals in Safari. Imported UseDisclosureReturn to define types for the infoModal and updateModal properties. Modified the arguments of the RenewalLikelihood and RenewalForecast components and updated their internal click event handlers to accommodate these new props.