(fix|cos-619): Fix for scrollbar overflowing About panel modals (#4221)
* (fix|cos-619): Fix for scrollbar overflowing About panel modals

Modify disclosure hook usage for RenewalForecast and RenewalLikelihood
Updated useDisclosure hook, moved it to an upper scope in the AccountPanel component to fix an issue with scrolling. The issue was that scrollbars were overlapping with renewal modals in Safari. Imported UseDisclosureReturn to define types for the infoModal and updateModal properties. Modified the arguments of the RenewalLikelihood and RenewalForecast components and updated their internal click event handlers to accommodate these new props.

* Remove optional chaining in AccountPanel modal conditions

Optional chaining has been removed from the isModalOpen logic in the AccountPanel component. This change was made because the properties being checked are guaranteed to be present in the objects, so there is no need for optional chaining. This simplifies the logic and may improve JavaScript runtime performance.

* Add id to useDisclosure in AccountPanel

In the AccountPanel component, id's have been added to the useDisclosure calls for the renewalLikelihood and renewalForecast modals. This change was made to avoid an overlapping scrollbar error in Safari. The link to the relevant issue on linear.app is given in the source comment.