(feat|COS-533): Add information is issue was created or updated to last touchpoint cell (#4226)
* (feat|COS-533): Add information is issue was created or updated to last touchpoint cell

This update introduces a feature that compares timestamp of issue creation and last update to determine whether an issue was created or updated. This functionality is implemented in `LastTouchpointCell.tsx` using a new utils function `diffetenceInMins` that computes the difference in minutes between two date instances.

Additionally, `createdAt` and `updatedAt` fields were added to 'Issue' in `getOrganizations.generated.ts` and `getOrganizations.graphql` to support this feature. Minor adjustments were also made to the graphql schema in `graphql.types.ts`.

* Update method name

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>

* Correct typo in DateTimeUtils function call

---------

Co-authored-by: coderabbitai[bot] <136622811+coderabbitai[bot]@users.noreply.github.com>