(fix|COS-712): Fixed placement of floating link menu, improved styling (#4316)
This commit rectifies the improper positioning of the floating link menu in the rich text editor. An adjustment was made in the react hook, using `useRef` to control the positioning of the element relative to its parent. This enables the menu to adjust its position dynamically thus avoiding overflow issues. CSS changes were also done to accommodate this position adjustment. The menu now places itself dynamically as per the available space.