(fix|COS-807): Fix unknown value of CC value in EmailStub (ui only solution) (#4420)
Added BCC field and optimize computation in EmailStub
The commit introduces the BCC field to the emailStub file to provide more information for each email. It also optimizes the computation of the 'to', 'cc', and 'bcc' variables by introducing the 'useMemo' React hook which helps prevent unnecessary computations by only executing when the 'email?.sentTo' array changes. This should help improve the performance of the component.