(fix|COS-597): Fix AboutPanel.tsx not saving information issue
Refactored the way organizations are saved in the 'About' panel to use debouncing. The previous implementation used setTimeout which was cleared before every new call. Now, the lodash debounce function is utilized instead. This prevents unnecessary multiple mutations when the data in the form is changed rapidly and most importantly we can safely flush data on back navigation