(feat|587): Add new OneTimeServiceModal component
This commit introduces a new component, OneTimeServiceModal, for adding one-time services to an organization's AccountPanel. This enhancement provides a more interactive way for users to include non-recurring services to an account. Changes also include improvements to some UI elements like updating the color of DotSingle icon for better thematic consistency, and better use of the readOnly state for the SyncSelect component to enhance user experience.