WIP: (feat|COS-855): Update AccountPanel component, add Contract DTO and ContractCard
Updated AccountPanel component to include ContractCard and ARRForecast components instead of RenewalForecast, TimeToRenewal and BillingDetailsCard. The change also includes the addition of Contract.dto.ts and ContractCard.tsx files corresponding to the new Contract module. The modals related to hooks 'useDisclosure' are commented out for future modification and connection. This change is part of the requirement to transition the organization's accounts to a contract-based system.