(feat|COS-855): Draft contract UI (#4437)
* WIP: (feat|COS-855): Update AccountPanel component, add Contract DTO and ContractCard

Updated AccountPanel component to include ContractCard and ARRForecast components instead of RenewalForecast, TimeToRenewal and BillingDetailsCard. The change also includes the addition of Contract.dto.ts and ContractCard.tsx files corresponding to the new Contract module. The modals related to hooks 'useDisclosure' are commented out for future modification and connection. This change is part of the requirement to transition the organization's accounts to a contract-based system.

* WIP: Added EmptyContracts view for accounts without contracts

This commit introduces a new EmptyContracts component that will be displayed within the AccountPanel when there are no contracts associated with the account. The component includes a button to prompt the creation of a new contract.

* Add ARRForecast component to AccountPanel

Added a new component named ARRForecast to the AccountPanel. This component displays Annual Recurring Revenue (ARR) forecast.

* Changed icon in EmptyContracts component

* Updated components to manage Contracts and Services

In this commit, a ServicesList component was added to allow the listing of services. Modifications were also made on the SubscriptionServiceModal, ContractCard, and ARRCost components to enable adding and updating of services in a Contract.

* Additionally, a one-time service component has been added, the SubscriptionServiceModal has been renamed to OneTimeServiceModal and components to handle one-time and subscription services were extracted out from the old SubscriptionServiceModal into independent forms. This was done to offer better structured, more reusable and more testable components. Also, FormInput component was modified to allow for custom FormLabelProps, providing more flexibility in label customization.

SubscriptionServiceModal was earlier tightly coupled, and handled both one-time and subscription services. By having separate forms, it provides a clear distinction between the two types of service and allows more flexibility and control. Also, renaming SubscriptionServiceModal to ServiceModal makes it clear that it can encompass any kind of service, and not just subscription ones.