(feat|COS-908): Implement feedback on ui
This commit updates the input fields related to the contract renewal cycle within the ContractCard component. Replacing a DatePicker with a custom FormSelect component with predefined options allows users to select from predetermined billing cycles, increasing usability and data consistency. Furthermore, the definition and usage of the billing frequency options are added to the AccountPanel utils file. Finally, the ContractDTO, SubscriptionServiceFrom and OneTimeServiceForm files have been updated to embody the changes.