feat:(COS-934) Contracts in customerOS (#4517)
* (feat|COS-855): Draft contract UI (#4437)

* WIP: (feat|COS-855): Update AccountPanel component, add Contract DTO and ContractCard

Updated AccountPanel component to include ContractCard and ARRForecast components instead of RenewalForecast, TimeToRenewal and BillingDetailsCard. The change also includes the addition of Contract.dto.ts and ContractCard.tsx files corresponding to the new Contract module. The modals related to hooks 'useDisclosure' are commented out for future modification and connection. This change is part of the requirement to transition the organization's accounts to a contract-based system.

* WIP: Added EmptyContracts view for accounts without contracts

This commit introduces a new EmptyContracts component that will be displayed within the AccountPanel when there are no contracts associated with the account. The component includes a button to prompt the creation of a new contract.

* Add ARRForecast component to AccountPanel

Added a new component named ARRForecast to the AccountPanel. This component displays Annual Recurring Revenue (ARR) forecast.

* Changed icon in EmptyContracts component

* Updated components to manage Contracts and Services

In this commit, a ServicesList component was added to allow the listing of services. Modifications were also made on the SubscriptionServiceModal, ContractCard, and ARRCost components to enable adding and updating of services in a Contract.

* Additionally, a one-time service component has been added, the SubscriptionServiceModal has been renamed to OneTimeServiceModal and components to handle one-time and subscription services were extracted out from the old SubscriptionServiceModal into independent forms. This was done to offer better structured, more reusable and more testable components. Also, FormInput component was modified to allow for custom FormLabelProps, providing more flexibility in label customization.

SubscriptionServiceModal was earlier tightly coupled, and handled both one-time and subscription services. By having separate forms, it provides a clear distinction between the two types of service and allows more flexibility and control. Also, renaming SubscriptionServiceModal to ServiceModal makes it clear that it can encompass any kind of service, and not just subscription ones.

* (feat|COS-902) Integrate with 'contract_Create' mutation + added optimistic update for contract and action timeline event (#4452)

* (feat|COS-891): Implement ui for timeline events for contract and service change events. (#4454)

Added preview modals for newly created timeline stubs

* (feat|COS-908): Implement feedback on ui (#4456)

This commit updates the input fields related to the contract renewal cycle within the ContractCard component. Replacing a DatePicker with a custom FormSelect component with predefined options allows users to select from predetermined billing cycles, increasing usability and data consistency. Furthermore, the definition and usage of the billing frequency options are added to the AccountPanel utils file. Finally, the ContractDTO, SubscriptionServiceFrom and OneTimeServiceForm files have been updated to embody the changes.

* (feat|COS-890): Implemented contract status select UI (#4458)

* (feat|COS-890): Implemented contract status select UI

* Reuse styles for account lifecycle select and contract status select

* Refactor ContractStatusSelect and add ConfirmationModals

* (feat|COS-907) (feat|COS-914) (feat|COS-854): Integrated with getContract query and createService mutation & Added missing ui parts for live contract ui (#4480)

* Added Contract Renewal card and introduce it in contract UI

* (feat|COS-907) (feat|COS-914): Integrated with getContract query and createService mutation
(feat|COS-854): Added missing ui parts for live contract ui

* Regenerate types after rebase

* (feat|COS-934): Integrated with, services and contract queries and mutations (#4513)

* (feat|COS-934): Integration with services list, ui improvements

* Fix createService form

* Fix calculateNextRenewalDate function