(fix|COS-984) (fix|COS-983) : Fix renewal date calculation, ui improvements (#4519)
* Added FileCheck02 icon and update ContractUrl input behavior

This commit introduces a new icon, FileCheck02, and updates the behavior of the ContractUrl input field in the AccountPanel component. Now, the FileCheck02 icon is displayed when a url is present. The handleUpdateContractUrl function was also updated to take an explicit "value" argument, which simplifies the function and removes the need for a separate "handleCancel" function. This change improves the clarity and predictability of the system's reaction to user inputs in the ContractUrl field.

* (fix|COS-984): Refine contract renewal logic, design updates and fix mutations

This commit includes several changes across contract management functionalities. For the contract renewal, the logic for Monthly and Annual renewals were adjusted ensuring the nextRenewal never points to a date in the past. For design updates, predefined borderRadius values were replaced by specific pixel definitions and added a pointer cursor on service cards for better UX. Also, the minimum allowed value for licenses and price/license in SubscriptionServiceForm is increased from 0 to 1 preventing users from entering incorrect values. Lastly, mutation operations for UpdateContract have been commented out as part of a fix to be carried out later.