COS-981 Fix rebasing errors (#4542)
* Revert "Rebase fix (#4540)"

This reverts commit 1b5f70761c8482f209f56b4ca2848e657e5fcb0a.

* Revert "(feat|COS-981): Integrate with opportunities (#4537) (#4538)"

This reverts commit f5c28324f0130fe330a9bd8d9644df63b7ac4865.

* (fix|COS-909): Fix HtmlContentRenderer for zero-width non-joiners (#4462)

This commit modifies the HtmlContentRenderer to replace zero-width non-joiner characters from the html content string before processing it with linkifyHtml function. This ensures that the HTML contents render properly irrespective of the presence of zero-width non-joiner characters which may cause rendering issues in some cases.

* feat:(COS-934) Contracts in customerOS (#4517)

* (feat|COS-855): Draft contract UI (#4437)

* WIP: (feat|COS-855): Update AccountPanel component, add Contract DTO and ContractCard

Updated AccountPanel component to include ContractCard and ARRForecast components instead of RenewalForecast, TimeToRenewal and BillingDetailsCard. The change also includes the addition of Contract.dto.ts and ContractCard.tsx files corresponding to the new Contract module. The modals related to hooks 'useDisclosure' are commented out for future modification and connection. This change is part of the requirement to transition the organization's accounts to a contract-based system.

* WIP: Added EmptyContracts view for accounts without contracts

This commit introduces a new EmptyContracts component that will be displayed within the AccountPanel when there are no contracts associated with the account. The component includes a button to prompt the creation of a new contract.

* Add ARRForecast component to AccountPanel

Added a new component named ARRForecast to the AccountPanel. This component displays Annual Recurring Revenue (ARR) forecast.

* Changed icon in EmptyContracts component

* Updated components to manage Contracts and Services

In this commit, a ServicesList component was added to allow the listing of services. Modifications were also made on the SubscriptionServiceModal, ContractCard, and ARRCost components to enable adding and updating of services in a Contract.

* Additionally, a one-time service component has been added, the SubscriptionServiceModal has been renamed to OneTimeServiceModal and components to handle one-time and subscription services were extracted out from the old SubscriptionServiceModal into independent forms. This was done to offer better structured, more reusable and more testable components. Also, FormInput component was modified to allow for custom FormLabelProps, providing more flexibility in label customization.

SubscriptionServiceModal was earlier tightly coupled, and handled both one-time and subscription services. By having separate forms, it provides a clear distinction between the two types of service and allows more flexibility and control. Also, renaming SubscriptionServiceModal to ServiceModal makes it clear that it can encompass any kind of service, and not just subscription ones.

* (feat|COS-902) Integrate with 'contract_Create' mutation + added optimistic update for contract and action timeline event (#4452)

* (feat|COS-891): Implement ui for timeline events for contract and service change events. (#4454)

Added preview modals for newly created timeline stubs

* (feat|COS-908): Implement feedback on ui (#4456)

This commit updates the input fields related to the contract renewal cycle within the ContractCard component. Replacing a DatePicker with a custom FormSelect component with predefined options allows users to select from predetermined billing cycles, increasing usability and data consistency. Furthermore, the definition and usage of the billing frequency options are added to the AccountPanel utils file. Finally, the ContractDTO, SubscriptionServiceFrom and OneTimeServiceForm files have been updated to embody the changes.

* (feat|COS-890): Implemented contract status select UI (#4458)

* (feat|COS-890): Implemented contract status select UI

* Reuse styles for account lifecycle select and contract status select

* Refactor ContractStatusSelect and add ConfirmationModals

* (feat|COS-907) (feat|COS-914) (feat|COS-854): Integrated with getContract query and createService mutation & Added missing ui parts for live contract ui (#4480)

* Added Contract Renewal card and introduce it in contract UI

* (feat|COS-907) (feat|COS-914): Integrated with getContract query and createService mutation
(feat|COS-854): Added missing ui parts for live contract ui

* Regenerate types after rebase

* (feat|COS-934): Integrated with, services and contract queries and mutations (#4513)

* (feat|COS-934): Integration with services list, ui improvements

* Fix createService form

* Fix calculateNextRenewalDate function

* (fix|COS-984) (fix|COS-983) : Fix renewal date calculation, ui improvements (#4519)

* Added FileCheck02 icon and update ContractUrl input behavior

This commit introduces a new icon, FileCheck02, and updates the behavior of the ContractUrl input field in the AccountPanel component. Now, the FileCheck02 icon is displayed when a url is present. The handleUpdateContractUrl function was also updated to take an explicit "value" argument, which simplifies the function and removes the need for a separate "handleCancel" function. This change improves the clarity and predictability of the system's reaction to user inputs in the ContractUrl field.

* (fix|COS-984): Refine contract renewal logic, design updates and fix mutations

This commit includes several changes across contract management functionalities. For the contract renewal, the logic for Monthly and Annual renewals were adjusted ensuring the nextRenewal never points to a date in the past. For design updates, predefined borderRadius values were replaced by specific pixel definitions and added a pointer cursor on service cards for better UX. Also, the minimum allowed value for licenses and price/license in SubscriptionServiceForm is increased from 0 to 1 preventing users from entering incorrect values. Lastly, mutation operations for UpdateContract have been commented out as part of a fix to be carried out later.

* (feat|COS-981): Integrate with opportunities

* Add Opportunity type to contracts in AccountPanel

In AccountPanel component, the Opportunity type was added to the imported types from GraphQL and applied to contract opportunities. This will ensure proper type checking and autocompletion for better code maintainability and fewer runtime errors.

* Replace `getFeatureIconColor` with `getARRColor` across all instances

This commit substitutes the method `getFeatureIconColor` with `getARRColor` in all the concerned files, including renewal-forecast and renewal-likelihood components. This change is introduced to improve the visual consistency on the application, attributing colors according to the ARR value associated with each event, increasing the user's understanding of these events.