(fix|COS-1032) Opportunities release bug fixes (#4590)
* Refactor formatCurrency and ContractStatus check; improve UI text clarity

Simplified the formatCurrency function in getFormattedCurrencyNumber.ts, removing abbreviation logic for cleaner readability of larger numbers. Also, set maximumFractionDigits to 2 for consistency in currency values display.
Adjusted the ContractStatus check in ContractSubtitle.tsx to base on status, making contract end status representation more accurate.

Additionally, this commit improves the clarity of tooltips, placeholder texts, messages across the Contract management components and made certain UI simplifications such as conditional checks in RenewalDetailsModal.tsx and ContractCard.tsx.

* Refactor UI alignment and enhance error messaging; improve date tracking in contracts

Refactored the user interface elements in RenewalARRCard.tsx to adjust the alignment of visual components more accurately for better UX. Changed the condition in the ContractSubtitle file from 'signedAt' to 'serviceStartedAt' for unified and improved date tracking in contract management.

Error messaging in the CreateServiceModal file was updated to be more relevant and providing the user with clear information. Introduced debounce and timeout adjustments in the ContractCard file.

* Update label name from 'Price/license' to 'Price/qty'

* Comment out "is retroactive version" checkbox until api is available