(fix|COS-1075): Improve the contract management experience and fix ui related issues.
Firstly, the previously commented out mutation code for updating contracts was reinstated, enabling optimistic updates for contract changes and toast error handling if the update fails.

Secondly, the name generation of a new contract was updated to default to 'Unnamed's contract' if there's no organization name available.

Thirdly, the contract renewal information is now correctly shown, hiding it if contract is ended or if there's no renewal cycle defined.

Lastly, in 'RenewalARRCard', a separate state control was added to handle the opening of the RenewalDetailsModal, ensuring the modal only opens when it's being acted upon in the corresponding contract context.