(fix|COS-1133):Refine currency format and service pricing in Spaces app
Improved the flexibility of the currency formatting function to support variable decimal places, allowing to depict prices more accurately. Lowered the minimum allowed value in the price/quantity fields in the service modals to permit smaller pricing values. This enhances price representation, especially for low-cost resources or granular billing. Additionally, changed the structure of 'opportunities' in the CreateServiceModal to correct the price reflecting and amended the title of 'renewalLikelihood' column for better semantic accuracy. Fixed column alignment