(fix|COS-1156): The 'timeAgo' method in DateTimeUtils has been updated to provide a simple 'today' & (fix|COS-1135): Update label on ContractCard (#4705)
* (fix|COS-1156): The 'timeAgo' method in DateTimeUtils has been updated to provide a simple 'today'
if the input date is the current date, thereby enhancing readability for users.
This change also resulted in refactoring of the usage of this function in RenewalARRCard component.
Previously, 'isToday' check was done explicitly in this component, but with this change,
the check has been encapsulated within the function call itself.

* (fix|COS-1135): Update label on ContractCard

Changed the label of the FormSelect in ContractCard from 'Contract renews' to 'Renewal cycle' for better clarification and user experience. The updated label more accurately reflects the selection purpose.